feat: acclimation actor

+++ src/main/scala/Acclimation/Extensions.scala
@@ -10,18 +10,18 @@ object Extensions:
     extension (p: Player)(using accl: Storage)
         def temperatureAcclimation =
             accl.getTemperature(p.getUniqueId())
-        def setTemperatureAcclimation(to: Float) =
+        def setTemperatureAcclimation(to: Double) =
             accl.setTemperature(p.getUniqueId(), to)
 
         def elevationAcclimation =
             accl.getElevation(p.getUniqueId())
-        def setElevationAcclimation(to: Float) =
+        def setElevationAcclimation(to: Double) =
             accl.setElevation(p.getUniqueId(), to)
         def longitudeAcclimation =
             accl.getLongitude(p.getUniqueId())
-        def setLongitudeAcclimation(to: Float) =
+        def setLongitudeAcclimation(to: Double) =
             accl.setLongitude(p.getUniqueId(), to)
         def latitudeAcclimation =
             accl.getLatitude(p.getUniqueId())
-        def setLatitudeAcclimation(to: Float) =
+        def setLatitudeAcclimation(to: Double) =
             accl.setLatitude(p.getUniqueId(), to)
