+++ src/main/scala/Acclimation/Storage.scala
@@ -10,33 +10,33 @@ import java.util.UUID
 import scala.language.implicitConversions
 
 class Storage()(using kvs: KeyVal):
-    private def get(player: UUID, key: String, default: Float): Float =
-        kvs.get[Float](player, key) match
+    private def get(player: UUID, key: String, default: Double): Double =
+        kvs.get[Double](player, key) match
             case Some(value) => value
             case None => default
-    private def set(player: UUID, key: String, value: Float) =
+    private def set(player: UUID, key: String, value: Double) =
         kvs.set(player, key, value)
 
     /** temperature ranges from 0.0 (coldest climate) to 1.0 (hottest climate) */
-    def getTemperature(player: UUID): Float =
+    def getTemperature(player: UUID): Double =
         get(player, "acclimation.temperature", 0.5)
-    def setTemperature(player: UUID, value: Float) =
+    def setTemperature(player: UUID, value: Double) =
         set(player, "acclimation.temperature", value)
 
     /** elevation ranges from -1.0 (bedrock) to 0.0 (sea level) to 1.0 (build limit) */
-    def getElevation(player: UUID): Float =
+    def getElevation(player: UUID): Double =
         get(player, "acclimation.elevation", 0.0)
-    def setElevation(player: UUID, value: Float) =
+    def setElevation(player: UUID, value: Double) =
         set(player, "acclimation.elevation", value)
 
     /** elevation ranges from -1.0 (westmost) to 0.0 (centre) to 1.0 (eastmost) */
-    def getLongitude(player: UUID): Float =
+    def getLongitude(player: UUID): Double =
         get(player, "acclimation.longitude", 0.0)
-    def setLongitude(player: UUID, value: Float) =
+    def setLongitude(player: UUID, value: Double) =
         set(player, "acclimation.longitude", value)
 
     /** latitude ranges from -1.0 (northmost) to 0.0 (centre) to 1.0 (southmost) */
-    def getLatitude(player: UUID): Float =
+    def getLatitude(player: UUID): Double =
         get(player, "acclimation.latitude", 0.0)
-    def setLatitude(player: UUID, value: Float) =
+    def setLatitude(player: UUID, value: Double) =
         set(player, "acclimation.latitude", value)
